Automatically assign current authenticated user to post when creating post

+++ src/main/java/com/social/socialapi/posts/PostService.java
@@ -2,11 +2,14 @@ package com.social.socialapi.posts;
 
 import java.util.List;
 
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import com.social.socialapi.user.User;
 import com.social.socialapi.user.UserRepository;
+import com.social.socialapi.utils.authUtils;
 
+import io.jsonwebtoken.security.SecurityException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
@@ -26,9 +29,11 @@ public class PostService {
         return post;
     }
 
-    public void createPost(Integer user_id, String body) {
+    public void createPost(String body) {
         Post post = new Post(body);
-        User user = userRepository.findById(user_id).orElseThrow();
+        String userEmail = authUtils.getAuthenticatedUser().getUsername();
+        User user = userRepository.findByEmail(userEmail).orElseThrow();
+
         post.setUser(user);
 
         postRepository.save(post);
@@ -38,6 +43,12 @@ public class PostService {
     public void updatePost(Integer id, String body) {
         Post post = postRepository.findById(id).orElseThrow();
 
+        UserDetails user = authUtils.getAuthenticatedUser();
+
+        if (!post.getUser().getEmail().equals(user.getUsername())) {
+            throw new SecurityException("You can't update this post");
+        }
+
         if (body != null) {
             post.setBody(body);
         }
@@ -46,6 +57,14 @@ public class PostService {
     }
 
     public void deletePost(Integer id) {
+        Post post = postRepository.findById(id).orElseThrow();
+
+        UserDetails user = authUtils.getAuthenticatedUser();
+
+        if (!post.getUser().getEmail().equals(user.getUsername())) {
+            throw new SecurityException("You can't update this post");
+        }
+
         postRepository.deleteById(id);
     }
 }
